moves new types into separate files
This also removes the dependency of the modules on MPI
