+++ include/ginkgo/core/base/lin_op.hpp
@@ -916,71 +916,6 @@ protected:
 };
 
 
-#if GINKGO_BUILD_MPI
-
-
-namespace experimental {
-
-
-/**
- * This mixin does the same as EnableLinOp, but for concrete
- * types that are derived from distributed::DistributedBase.
- *
- * @see EnableLinOp.
- *
- * @tparam ConcreteLinOp  the concrete LinOp which is being implemented that
- *                        is derived from distributed::DistributedBase
- *                        [CRTP parameter]
- * @tparam PolymorphicBase  parent of ConcreteLinOp in the polymorphic
- *                          hierarchy, has to be a subclass of LinOp
- *
- * @ingroup LinOp
- */
-template <typename ConcreteLinOp, typename PolymorphicBase = LinOp>
-class EnableDistributedLinOp
-    : public EnableDistributedPolymorphicObject<ConcreteLinOp, PolymorphicBase>,
-      public EnablePolymorphicAssignment<ConcreteLinOp> {
-public:
-    using EnableDistributedPolymorphicObject<
-        ConcreteLinOp, PolymorphicBase>::EnableDistributedPolymorphicObject;
-
-    const ConcreteLinOp* apply(const LinOp* b, LinOp* x) const
-    {
-        PolymorphicBase::apply(b, x);
-        return self();
-    }
-
-    ConcreteLinOp* apply(const LinOp* b, LinOp* x)
-    {
-        PolymorphicBase::apply(b, x);
-        return self();
-    }
-
-    const ConcreteLinOp* apply(const LinOp* alpha, const LinOp* b,
-                               const LinOp* beta, LinOp* x) const
-    {
-        PolymorphicBase::apply(alpha, b, beta, x);
-        return self();
-    }
-
-    ConcreteLinOp* apply(const LinOp* alpha, const LinOp* b, const LinOp* beta,
-                         LinOp* x)
-    {
-        PolymorphicBase::apply(alpha, b, beta, x);
-        return self();
-    }
-
-protected:
-    GKO_ENABLE_SELF(ConcreteLinOp);
-};
-
-
-}  // namespace experimental
-
-
-#endif
-
-
 /**
  * This is an alias for the EnableDefaultFactory mixin, which correctly sets the
  * template parameters to enable a subclass of LinOpFactory.
